send special exit code when exec is killed by OOM-killer
1161 is a new error code but because PAE status is already int,
this does not change the wire protocol per se.

When talking to old runv, runv would trim the exit code to uint8,
getting the original process exit code.

+++ src/exec.h
@@ -21,7 +21,7 @@ struct hyper_exec {
 	int			init;
 	int			ptyfd;
 	uint8_t			close_stdin_request;
-	uint8_t			code;
+	uint32_t		code;
 	uint8_t			exit;
 	uint8_t			ref;
